Add documentation to the project
Documentation is hosted on readthedocs.org, it’s free for public
open-source project and a popular choice.
The Sphinx readthedocs theme was picked for the same reasons.

The code examples are verified with doctest.

Most documentation for the fields live in the code directly, in an
attempt to keep them in sync.